Add raw module
This can be used in places where you need to integrate allocation with existing infrastructure (like in std), and thus do not want extra abstraction.
