+++ src/lib.rs
@@ -26,3 +26,4 @@
 
 pub mod alloc;
 pub mod global;
+pub mod raw;
